python/ommx/pyproject.toml
[build-system]
requires = ["maturin >= 1.12.2"]
build-backend = "maturin"

[project]
name = "ommx"

version = "3.0.0a3"
description = "Open Mathematical prograMming eXchange (OMMX)"
authors = [{ name = "Jij Inc.", email = "info@j-ij.com" }]
readme = "README.md"

classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
  "License :: OSI Approved :: Apache Software License",
  "License :: OSI Approved :: MIT License",
]

requires-python = ">=3.10"
dependencies = [
  "numpy>=1.23.0, <3.0.0",
  "pandas>=2.0.0, <3.0.0",
  "pyarrow>=16.0.0, <23.0.0",
  "python-dateutil>=2.9.0, <3.0.0",
  "typing-extensions >= 4.12.2, < 5.0.0",
]

[project.optional-dependencies]
# ``%%ommx_trace`` and ``ommx.tracing`` require the OpenTelemetry Python SDK
# (for the ``TracerProvider`` / ``SpanProcessor`` types). Users who do not
# need tracing should not have to pay that cost, hence an optional extra:
# ``pip install ommx[tracing]``.
tracing = [
  "opentelemetry-sdk>=1.20.0",
  "ipython>=7.0.0",
]

[project.urls]
Repository = "https://github.com/Jij-Inc/ommx"
Issues = "https://github.com/Jij-Inc/ommx/issues"

[tool.maturin]
module-name = "ommx._ommx_rust"
features = ["pyo3/extension-module", "pyo3/abi3-py310"]
python-source = "."

[tool.cibuildwheel]
build = "cp313-pyodide_wasm32"

[tool.cibuildwheel.pyodide]
pyodide-version = "0.29.3"
environment = { RUSTUP_TOOLCHAIN = "nightly-2026-01-01" }
before-build = "rustup toolchain install nightly-2026-01-01 && rustup target add wasm32-unknown-emscripten --toolchain nightly-2026-01-01"
config-settings = { build-args = "--no-default-features" }

[tool.pyo3-stub-gen]
generate-init-py = ["ommx.v1", "ommx.artifact"]

[tool.pyo3-stub-gen.doc-gen]
output-dir = "../../docs/api"
json-output = "api_reference.json"
generate-index = false
contents-table = true
separate-items = true

[tool.uv]
cache-keys = [
  { file = "pyproject.toml" },
  { file = "src/**/*.rs" },
  # Python SDK should be rebuilt when Rust SDK is updated.
  { file = "../../Cargo.lock" },
  { file = "../../rust/ommx/src/**/*.rs" },
]
